Add files via upload
These are the files for the graph and heap implementation, as well as the JSON file library needed to take in the data. Unfortunately, github is not allowing me to upload the word map file with all the data since it is so large, so hopefully we can find a way to do that.